rename storage module
